+++ src/types.jl
@@ -527,6 +527,7 @@ const INFERENCE_UNDEF_STATIC_PARAM_CODE = "inference/undef-static-param" # curre
 const INFERENCE_FIELD_ERROR_CODE = "inference/field-error"
 const INFERENCE_BOUNDS_ERROR_CODE = "inference/bounds-error"
 const INFERENCE_METHOD_ERROR_CODE = "inference/method-error"
+const INFERENCE_UNDEF_KEYWORD_CODE = "inference/undef-keyword"
 const INFERENCE_TYPE_ERROR_NON_BOOL_COND_CODE = "inference/type-error/non-bool-cond"
 const INFERENCE_TYPE_ERROR_TYPE_ASSERT_CODE = "inference/type-error/type-assert"
 const DEPRECATED_INFERENCE_NON_BOOLEAN_COND_CODE = "inference/non-boolean-cond"
@@ -557,6 +558,7 @@ const ALL_DIAGNOSTIC_CODES = Set{String}(String[
     INFERENCE_FIELD_ERROR_CODE,
     INFERENCE_BOUNDS_ERROR_CODE,
     INFERENCE_METHOD_ERROR_CODE,
+    INFERENCE_UNDEF_KEYWORD_CODE,
     INFERENCE_TYPE_ERROR_NON_BOOL_COND_CODE,
     INFERENCE_TYPE_ERROR_TYPE_ASSERT_CODE,
     TESTRUNNER_TEST_FAILURE_CODE,